Closes #1479 Active-Passive data sync

+++ clustered/common/src/main/java/org/ehcache/clustered/common/internal/messages/ClientIDTrackerMessage.java
@@ -119,8 +119,8 @@ public class ClientIDTrackerMessage extends EhcacheEntityMessage {
     }
 
     @Override
-    public int concurrencyKey() {
-      return (int) (this.cacheId.hashCode() + key);
+    public long concurrencyKey() {
+      return (this.cacheId.hashCode() + key);
     }
   }
 }
